+++ changelog.md
@@ -1,3 +1,6 @@
+# Upcoming
+* Fixed Respiration (oxygen_bonus attribute) not working
+
 # 5.6.3.0
 * Tiredness
   * Time skipped on sleep with Tired I is now 9h instead of 12h (still +3h per higher level of Tired)